@@ -7,14 +7,12 @@ import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.MockitoJUnitRunner;
 
-@RunWith(MockitoJUnitRunner.class)
 public class NthRootCalculatorUnitTest {
 
-    @InjectMocks
-    private NthRootCalculator nthRootCalculator;
+    private NthRootCalculator nthRootCalculator = new NthRootCalculator();
 
     @Test
-    public void giventThatTheBaseIs125_andTheExpIs3_whenCalculateIsCalled_thenTheResultIsTheCorrectOne() {
+    public void whenBaseIs125AndNIs3_thenNthRootIs5() {
        Double result = nthRootCalculator.calculate(125.0, 3.0);
        assertEquals(result, (Double) 5.0d);
     }
