@@ -8,8 +8,8 @@ import org.springframework.http.server.reactive.HttpHandler;
 import org.springframework.http.server.reactive.ReactorHttpHandlerAdapter;
 import org.springframework.web.reactive.config.EnableWebFlux;
 import org.springframework.web.server.adapter.WebHttpHandlerBuilder;
-import reactor.ipc.netty.NettyContext;
-import reactor.ipc.netty.http.server.HttpServer;
+import reactor.netty.DisposableServer;
+import reactor.netty.http.server.HttpServer;
 
 @ComponentScan(basePackages = {"com.baeldung.reactive.security"})
 @EnableWebFlux
@@ -18,17 +18,16 @@ public class SpringSecurity5Application {
     public static void main(String[] args) {
         try (AnnotationConfigApplicationContext context =
                      new AnnotationConfigApplicationContext(SpringSecurity5Application.class)) {
-            context.getBean(NettyContext.class).onClose().block();
+            context.getBean(DisposableServer.class).onDispose().block();
         }
     }
 
     @Bean
-    public NettyContext nettyContext(ApplicationContext context) {
+    public DisposableServer nettyContext(ApplicationContext context) {
         HttpHandler handler = WebHttpHandlerBuilder.applicationContext(context)
                 .build();
         ReactorHttpHandlerAdapter adapter = new ReactorHttpHandlerAdapter(handler);
-        HttpServer httpServer = HttpServer.create("localhost", 8080);
-        return httpServer.newHandler(adapter).block();
+        return HttpServer.create().host("localhost").port(8080).handle(adapter).bind().block();
     }
 
 }
