@@ -1,15 +1,16 @@
 package org.baeldung.boot.config;
 
+import java.util.List;
+
 import org.baeldung.boot.converter.GenericBigDecimalConverter;
+import org.baeldung.boot.converter.StringToAbstractEntityConverterFactory;
 import org.baeldung.boot.converter.StringToEmployeeConverter;
-import org.baeldung.boot.converter.StringToEnumConverterFactory;
+import org.baeldung.boot.converter.StringToEnumConverter;
 import org.baeldung.boot.web.resolver.HeaderVersionArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
@@ -22,7 +23,8 @@ public class WebConfig implements WebMvcConfigurer {
     @Override
     public void addFormatters(FormatterRegistry registry) {
         registry.addConverter(new StringToEmployeeConverter());
-        registry.addConverterFactory(new StringToEnumConverterFactory());
+        registry.addConverter(new StringToEnumConverter());
+        registry.addConverterFactory(new StringToAbstractEntityConverterFactory());
         registry.addConverter(new GenericBigDecimalConverter());
     }
 }
