@@ -1,8 +1,10 @@
 package com.baeldung.spring.cloud.feign.client;
 
-import com.baeldung.spring.cloud.eureka.client.GreetingController;
-import org.springframework.cloud.netflix.feign.FeignClient;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 @FeignClient("spring-cloud-eureka-client")
-public interface GreetingClient extends GreetingController {
+public interface GreetingClient {
+    @RequestMapping("/greeting")
+    String greeting();
 }
