@@ -1,17 +1,21 @@
 package com.baeldung.config;
 
+import com.baeldung.models.WebsiteUser;
+import com.baeldung.projections.CustomBook;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.rest.core.config.RepositoryRestConfiguration;
-import org.springframework.data.rest.webmvc.config.RepositoryRestConfigurerAdapter;
-
-import com.baeldung.projections.CustomBook;
-
+import org.springframework.data.rest.core.mapping.ExposureConfiguration;
+import org.springframework.data.rest.webmvc.config.RepositoryRestConfigurer;
+import org.springframework.http.HttpMethod;
 
 @Configuration
-public class RestConfig extends RepositoryRestConfigurerAdapter{
+public class RestConfig implements RepositoryRestConfigurer {
 
     @Override
     public void configureRepositoryRestConfiguration(RepositoryRestConfiguration repositoryRestConfiguration){
         repositoryRestConfiguration.getProjectionConfiguration().addProjection(CustomBook.class);
+        ExposureConfiguration config = repositoryRestConfiguration.getExposureConfiguration();
+        config.forDomainType(WebsiteUser.class).withItemExposure((metadata, httpMethods) ->
+          httpMethods.disable(HttpMethod.PATCH));
     }
 }
