@@ -14,7 +14,7 @@ import org.springframework.http.converter.xml.MarshallingHttpMessageConverter;
 import org.springframework.oxm.xstream.XStreamMarshaller;
 import org.springframework.web.servlet.config.annotation.ContentNegotiationConfigurer;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.text.SimpleDateFormat;
 import java.util.List;
@@ -25,7 +25,7 @@ import java.util.List;
 @Configuration
 @EnableWebMvc
 @ComponentScan({ "org.baeldung.web" })
-public class WebConfig extends WebMvcConfigurerAdapter {
+public class WebConfig implements WebMvcConfigurer {
 
     public WebConfig() {
         super();
@@ -48,7 +48,6 @@ public class WebConfig extends WebMvcConfigurerAdapter {
         messageConverters.add(new ProtobufHttpMessageConverter());
         messageConverters.add(new KryoHttpMessageConverter());
         messageConverters.add(new StringHttpMessageConverter());
-        super.configureMessageConverters(messageConverters);
     }
 
     private HttpMessageConverter<Object> createXmlHttpMessageConverter() {
