@@ -3,9 +3,7 @@ package com.baeldung.junit5.mockito;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -16,6 +14,7 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.invocation.InvocationOnMock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.stubbing.Answer;
 
 import com.baeldung.junit5.mockito.repository.MailClient;
@@ -25,21 +24,24 @@ import com.baeldung.junit5.mockito.service.DefaultUserService;
 import com.baeldung.junit5.mockito.service.Errors;
 import com.baeldung.junit5.mockito.service.UserService;
 
-@RunWith(JUnitPlatform.class)
 @ExtendWith(MockitoExtension.class)
+@RunWith(JUnitPlatform.class)
 public class UserServiceUnitTest {
     
     UserService userService;
+    SettingRepository settingRepository;
     @Mock UserRepository userRepository;
-    
+    @Mock MailClient mailClient;
+
     User user;
     
     @BeforeEach
-    void init(@Mock SettingRepository settingRepository, @Mock MailClient mailClient) {
+    void init(@Mock SettingRepository settingRepository) {
         userService = new DefaultUserService(userRepository, settingRepository, mailClient);
-        when(settingRepository.getUserMinAge()).thenReturn(10);
+        lenient().when(settingRepository.getUserMinAge()).thenReturn(10);
         when(settingRepository.getUserNameMinLength()).thenReturn(4);
-        when(userRepository.isUsernameAlreadyExists(any(String.class))).thenReturn(false);
+        lenient().when(userRepository.isUsernameAlreadyExists(any(String.class))).thenReturn(false);
+        this.settingRepository = settingRepository;
     }
     
     @Test
@@ -56,7 +58,9 @@ public class UserServiceUnitTest {
                 return user;
             }
         });
-        
+
+        userService = new DefaultUserService(userRepository, settingRepository, mailClient);
+
         // When
         User insertedUser = userService.register(user);
         
